+++ ci/tasks/terraform/apply_terraform.sh
@@ -5,7 +5,9 @@
 set -euo pipefail
 
 : ${TERRAFORM_SOURCE}
+: ${TF_VAR_aws_account_id}
 : ${TF_VAR_environment}
+: ${TF_VAR_slack_alert_token}
 : ${TF_VAR_deployment_role}
 : ${WORKSPACE_KEY_INFIX}
 : ${AWS_DEFAULT_REGION}
@@ -13,7 +15,6 @@ set -euo pipefail
 : ${S3_KEY}
 : ${username}
 : ${password}
-
 
 echo "Setting netrc creds"
 rm -f $HOME/.netrc
@@ -27,13 +28,15 @@ terraform init \
     -backend-config "workspace_key_prefix=${WORKSPACE_KEY_INFIX}" \
     -backend-config "role_arn=${TF_VAR_deployment_role}"
 echo "starting terraform plan"
-if [ -z ${TF_WORKSPACE+x} ]; then export TF_WORKSPACE=`cat ../.git/resource/head_name | tr "[:upper:]" "[:lower:]"`; else echo "Workspace already set"; fi
+if [ -z ${TF_WORKSPACE+x} ]; then export TF_WORKSPACE=`cat ../.workspace | tr "[:upper:]" "[:lower:]"`; else echo "Workspace already set"; fi
 echo Workspace: ${TF_WORKSPACE}
 terraform plan -out=plan.tfstate
 echo "starting terraform apply"
 terraform apply \
     -auto-approve \
-    -var="environment=${TF_VAR_environment}"
+    -var="environment=${TF_VAR_environment}" \
+    -var="slack_alert_token=${TF_VAR_slack_alert_token}" \
+    -var="aws_account_id=${TF_VAR_aws_account_id}"
 rm plan.tfstate
 echo "done"
 echo "DEPLOY_URL=`terraform output -raw website_url`" > ../../GITHUB_OUTPUT/output.txt