+++ ci/tasks/terraform/destroy_terraform.sh
@@ -4,13 +4,12 @@ set -eu
 
 export HOME=$(pwd)
 
+: ${TF_VAR_aws_account_id}
+
 echo "machine github.com login ${CONCOURSE_ACCESS_TOKEN} password x-oauth-basic" > ~/.netrc
 chmod 600 ~/.netrc
 
-cd repo
-
 # --------------------------
-
 export TF_VAR_head_sha=$(cat .git/resource/head_sha)
 export TF_VAR_head_sha_short=$(cat .git/resource/head_sha_short)
 #
@@ -30,7 +29,7 @@ terraform init \
 
 # --------------------------
 
-export TF_WORKSPACE=`cat ../.git/resource/head_name | tr "[:upper:]" "[:lower:]"`
+export TF_WORKSPACE=`cat ../.workspace | tr "[:upper:]" "[:lower:]"`
 echo "Workspace: ${TF_WORKSPACE}"
 
 # --------------------------