+++ contentful-data/migrations/preprod-export.js
@@ -1,0 +1,3 @@
+function migrationFunction(migration, context) {
+}
+module.exports = migrationFunction;