+++ sml_builder/glossary.py
@@ -1,3 +1,4 @@
+import re
 from json import loads
 from os import listdir
 
@@ -6,7 +7,13 @@ from flask import render_template
 
 from sml_builder import app
 
-from .utils import _page_not_found, convert_term
+from .utils import _page_not_found
+
+
+@app.template_filter("convert_name")
+def convert_term(value):
+    converted = re.sub("[^0-9a-z]+", "-", value.lower())
+    return converted
 
 
 @app.route("/resources/glossary")
@@ -27,12 +34,14 @@ def display_glossary():
                     "term": glossary_term["term"],
                     "hash": convert_term(glossary_term["term"]),
                     "letter": letter_value,
-                    "related": glossary_term["related"]
-                    if "related" in glossary_term
-                    else [],
-                    "external_links": glossary_term["external_links"]
-                    if "external_links" in glossary_term
-                    else [],
+                    "related": (
+                        glossary_term["related"] if "related" in glossary_term else []
+                    ),
+                    "external_links": (
+                        glossary_term["external_links"]
+                        if "external_links" in glossary_term
+                        else []
+                    ),
                     "meaning": glossary_term["meaning"],
                 }
             )