+++ terraform/variables.tf
@@ -1,9 +1,19 @@
+variable "aws_account_id" {
+  type = string
+}
+
 variable "environment" {
   type = string
 }
+
+variable "slack_alert_token" {
+  type = string
+}
+
 variable "deployment_role" {
   type = string
 }
+
 locals {
   domain_name_base = {
     dev : "dev-sml.aws.onsdigital.uk"